revert refactor to checks/evaluation
