+++ checks/raw/binary_artifact.go
@@ -17,17 +17,34 @@ package raw
 import (
 	"fmt"
 	"path/filepath"
+	"regexp"
 	"strings"
 	"unicode"
 
+	semver "github.com/Masterminds/semver/v3"
 	"github.com/h2non/filetype"
 	"github.com/h2non/filetype/types"
+	"github.com/rhysd/actionlint"
 
 	"github.com/ossf/scorecard/v4/checker"
 	"github.com/ossf/scorecard/v4/checks/fileparser"
 	"github.com/ossf/scorecard/v4/clients"
 	sce "github.com/ossf/scorecard/v4/errors"
 )
+
+var (
+	gradleWrapperValidationActionRegex             = regexp.MustCompile(`^gradle\/wrapper-validation-action@v?(.+)$`)
+	gradleWrapperValidationActionVersionConstraint = mustParseConstraint(`>= 1.0.0`)
+)
+
+// mustParseConstraint attempts parse of semver constraint, panics if fail.
+func mustParseConstraint(c string) *semver.Constraints {
+	if c, err := semver.NewConstraint(c); err != nil {
+		panic(fmt.Errorf("failed to parse constraint: %w", err))
+	} else {
+		return c
+	}
+}
 
 // BinaryArtifacts retrieves the raw data for the Binary-Artifacts check.
 func BinaryArtifacts(c clients.RepoClient) (checker.BinaryArtifactData, error) {
@@ -40,6 +57,33 @@ func BinaryArtifacts(c clients.RepoClient) (checker.BinaryArtifactData, error) {
 		return checker.BinaryArtifactData{}, fmt.Errorf("%w", err)
 	}
 
+	// Check if gradle-wrapper.jar present
+	hasGradleWrappers := false
+	if len(files) > 0 {
+		for _, f := range files {
+			if filepath.Base(f.Path) == "gradle-wrapper.jar" {
+				hasGradleWrappers = true
+				break
+			}
+		}
+	}
+	if hasGradleWrappers {
+		// Gradle wrapper JARs present, so check that they are validated
+		if ok, err := gradleWrapperValidated(c); ok && err == nil {
+			// It has been confirmed that latest commit has validated JARs!
+			// Remove Gradle wrapper JARs from files.
+			filterFiles := []checker.File{}
+			for _, f := range files {
+				if filepath.Base(f.Path) != "gradle-wrapper.jar" {
+					filterFiles = append(filterFiles, f)
+				}
+			}
+			files = filterFiles
+		} else if err != nil {
+			return checker.BinaryArtifactData{Files: files}, fmt.Errorf(
+				"failure checking for Gradle wrapper validating Action: %w", err)
+		}
+	}
 	// No error, return the files.
 	return checker.BinaryArtifactData{Files: files}, nil
 }
@@ -126,4 +170,81 @@ func isText(content []byte) bool {
 		}
 	}
 	return true
+}
+
+// gradleWrapperValidated checks for the gradle-wrapper-verify action being
+// used in a non-failing workflow on the latest commit.
+func gradleWrapperValidated(c clients.RepoClient) (bool, error) {
+	gradleWrapperValidatingWorkflowFile := ""
+	err := fileparser.OnMatchingFileContentDo(c, fileparser.PathMatcher{
+		Pattern:       ".github/workflows/*",
+		CaseSensitive: false,
+	}, checkWorkflowValidatesGradleWrapper, &gradleWrapperValidatingWorkflowFile)
+	if err != nil {
+		return false, fmt.Errorf("%w", err)
+	}
+	if gradleWrapperValidatingWorkflowFile != "" {
+		// If validated, check that latest commit has a relevant successful run
+		runs, err := c.ListSuccessfulWorkflowRuns(gradleWrapperValidatingWorkflowFile)
+		if err != nil {
+			return false, fmt.Errorf("failure listing workflow runs: %w", err)
+		}
+		commits, err := c.ListCommits()
+		if err != nil {
+			return false, fmt.Errorf("failure listing commits: %w", err)
+		}
+		if len(commits) < 1 || len(runs) < 1 {
+			return false, nil
+		}
+		for _, r := range runs {
+			if *r.HeadSHA == commits[0].SHA {
+				// Commit has corresponding successful run!
+				return true, nil
+			}
+		}
+	}
+	return false, nil
+}
+
+// checkWorkflowValidatesGradleWrapper checks that the current workflow file
+// is indeed using the gradle/wrapper-validation-action action, else continues.
+func checkWorkflowValidatesGradleWrapper(path string, content []byte, args ...interface{}) (bool, error) {
+	validatingWorkflowFile, ok := args[0].(*string)
+	if !ok {
+		return false, fmt.Errorf("checkWorkflowValidatesGradleWrapper expects arg[0] of type *string: %w", errInvalidArgType)
+	}
+
+	action, errs := actionlint.Parse(content)
+	if len(errs) > 0 {
+		return true, errs[0]
+	}
+
+	for _, j := range action.Jobs {
+		for _, s := range j.Steps {
+			ea, ok := s.Exec.(*actionlint.ExecAction)
+			if !ok {
+				continue
+			}
+			if ea.Uses == nil {
+				continue
+			}
+			sms := gradleWrapperValidationActionRegex.FindStringSubmatch(ea.Uses.Value)
+			if len(sms) > 1 {
+				v, err := semver.NewVersion(sms[1])
+				if err != nil {
+					// Couldn't parse version, hopefully another step has
+					// a correct one.
+					continue
+				}
+				if !gradleWrapperValidationActionVersionConstraint.Check(v) {
+					// Version out of acceptable range.
+					continue
+				}
+				// OK! This is it.
+				*validatingWorkflowFile = filepath.Base(path)
+				return true, nil
+			}
+		}
+	}
+	return true, nil
 }
